requirements.txt
fastapi
uvicorn
pydantic
requests
gunicorn
python-dotenv
google-generativeai
